requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
Click==7.0
cryptography==2.7
dlib
face-recognition-models==0.3.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy
opencv-python
Pillow
pycparser==2.19
pyOpenSSL==19.0.0
six==1.12.0
Werkzeug==0.15.5
pandas
matplotlib==3.0.2
scikit-learn==0.21.2
cmake
face_recognition
pyautogui
seaborn
tensorflow==2.0.0b1
scikit-image
imutils
google-cloud-vision
utils
